requirements.txt
streamlit
openai
requests
Pillow
python-dotenv
ibm-watsonx-ai